requirements.txt
# requirements.txt
# Generated from .venv in this workspace on 2025-10-04
pandas==2.3.3
numpy==2.3.3
pytz==2025.2
tzdata==2025.2
python-dateutil==2.9.0.post0
ipykernel==6.30.1
matplotlib==3.10.6
seaborn==0.13.2
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.19
openpyxl>=3.1.2
Flask==2.3.3
